refactor: move answer endpoint and tests into own files
