+++ api/router/exam.py
@@ -1,9 +1,7 @@
 from fastapi import APIRouter, Depends, HTTPException
-from sqlmodel import Session, asc, select
 from starlette.status import HTTP_404_NOT_FOUND
 
-from api.dependencies import get_assessment, get_assessment_id, get_session
-from api.schemas.answer import Answer, AnswerRead
+from api.dependencies import get_assessment
 from api.schemas.exam import Assessment, AssessmentSummary, Question
 
 exam_router = APIRouter()
@@ -27,37 +25,6 @@ def get_question(
     if question_number not in assessment.questions:
         raise HTTPException(status_code=HTTP_404_NOT_FOUND, detail="Question not found")
     return assessment.questions[question_number]
-
-
-# ------------------------------------------ Answers
-
-
-@exam_router.get(
-    "/questions/{question_number}/answer",
-    tags=["answers"],
-    response_model=list[AnswerRead],
-    summary="Retrieve answer for question by user",
-    description="""
-Retrieve the given latest answer by a user including part, section, task.
-""",
-)
-def get_answer(
-    question_number: int,
-    session: Session = Depends(get_session),
-    assessment_id: str = Depends(get_assessment_id),
-):
-    query = (
-        select(Answer)
-        .where(
-            Answer.question == question_number,
-            Answer.exam_id == assessment_id,
-            Answer.username == "hpotter",
-        )
-        .order_by(asc(Answer.part), asc(Answer.section), asc(Answer.task))
-    )
-    answers = session.exec(query).all()
-
-    return answers
 
 
 @exam_router.get(
