+++ tests/api/test_exam.py
@@ -1,4 +1,3 @@
-from api.factories.exam import AnswerFactory
 from api.schemas.exam import TaskType
 
 
@@ -25,24 +24,6 @@ def test_can_get_question_for_exam(client):
         {"value": "a", "label": "Red pill"},
         {"value": "b", "label": "Blue pill"},
     ]
-
-
-def test_can_get_answer_for_question_by_user(web_client, answer_factory):
-    answer = answer_factory.create_batch(
-        size=3, exam_id="y2023_12345_exam", question=1, username="hpotter"
-    )
-
-    res = web_client.get("/questions/1/answer")
-    assert res.status_code == 200
-    assert len(res.json()) == 3
-
-    assert all(answer["question"] == 1 for answer in res.json())
-
-
-def test_can_get_empty_list_for_no_answers_to_known_exam(web_client):
-    res = web_client.get("/questions/1/answer")
-    assert res.status_code == 200
-    assert len(res.json()) == 0
 
 
 def test_404_in_case_of_missing_question(client):
